requirements.txt
camel-ai[all]==0.2.16
python-dotenv>=1.0.0
Pillow>=10.0.0 